Set LIBCLANG_PATH for bindgen to work
Yes, ".cargo" is in .gitignore and this is better set in the developer
shell's environment, but having it here documents the issue and greatly
helps hacking on it.

https://doc.rust-lang.org/nightly/cargo/reference/config.html#env
`[env]` has been stabalised in Rust 1.56.0, so crank the requirement.

This is probably best dealt with in the OpenBSD port's Makefile once
psst is ready for it.