Queue behavior & position tracking for consistency
Trying to improve the queue struct, meaning the way it behaves during playback. So that the order of songs in the queue is consistent when switching between different playback behaviors or modifying queue. 

This fixes the following error, which makes the program unusable until restarting it:

```
thread '<unnamed>' panicked at 'index out of bounds: the len is 2 but the index is 10', 
psst-core/src/player/queue.rs:57:13 
```

So now it's making sure that the positions of the songs in the queue are always tracked properly by initializing the positions vector with the original order of the items. So that the songs are always played in the right order even if the queue behavior changes. 

Also some typo (behaviour -> behavior) and removed some unnecessary code.